Fixed python script and add init script
This fixes the python script which was not working with OpenCV 4.5.4.
Also adds the read_meter.sh file which is the script to take a picture
and then read the data.  The init.sh file creates the cron job that
calls read_meter.sh every 5 minutes. Hope is this will facilitate
installation.